Suggest similar names to combat spelling errors
This will correct my little brother's mistakes in spelling names.
Ideally I'd like to make the suggested names hyperlinked
so you could just click on them to load but that's probably overkill.

This requries computing the Damerau–Levenshtein distance
between all the names, which seems to be pretty slow
and makes the requests much more expesnive.