Do string similarity processing client side
This should *significantly* reduce server load,
since we don't have to do the expensive damerau levenshtein
algorithim on all the data.
Unfortunately this means we have to send all the known names to
the client side, instead of only sending the 'similar' ones.
We were already doing this processing before the similarity testing,
so this should only increase bandwidth and not server CPU time.