Use rocket_contrib to pool database connections
Loading data is now lightning fast and we still use very
little memory.
Unfortunately, we had to downgrade the version of `rusqlite`
but that's not a big price to pay :)

From my tests on my laptop it uses at most 14 MB of memory.
This is in contrast to the 383 MB of memory it used to use.